プルリク修正
