--- nextjs-blog/pages/posts/first-post
@@ -1,3 +1,0 @@
-export default function FirstPost() {
-  return <h1>First Post</h1>
-}
