--- nextjs-blog/pages/user.js
@@ -1,8 +1,0 @@
-<script>
-    const app = new Vue({
-        el : '#app',
-        data: {
-            isOpen: false,
-        }
-    })
-</script>
